had to remove uint and ulong from lib/types.h due to conflict with other library. this snowballed into a massive search+destroy of the hodgepodge of mostly equivalent types we had in use (int, uint, unsigned, unsigned int, i32, u32, ulong, uintN).
it is more efficient to use 64-bit types in 64-bit mode, so the preferred default is size_t (for anything remotely resembling a size or index). tile coordinates are ssize_t to allow more efficient conversion to/from floating point. flags are int because we almost never need more than 15 distinct bits, bit test/set is not slower and int is fastest to type. finally, some data that is pretty much directly passed to OpenGL is now typed accordingly.

after several hours, the code now requires fewer casts and less guesswork.

other changes:
- unit and player IDs now have an "invalid id" constant in the respective class to avoid casting and -1
- fix some endian/64-bit bugs in the map (un)packing. added a convenience function to write/read a size_t.
- ia32: change CPUID interface to allow passing in ecx (required for cache topology detection, which I need at work). remove some unneeded functions from asm, replace with intrinsics where possible.

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@5942 3db68df2-c116-0410-a063-a993310a9797

+++ source/graphics/Camera.h
@@ -18,10 +18,10 @@ extern int g_mouse_x, g_mouse_y;
 // view port
 struct SViewPort
 {
-	unsigned int m_X;
-	unsigned int m_Y;
-	unsigned int m_Width;
-	unsigned int m_Height;
+	size_t m_X;
+	size_t m_Y;
+	size_t m_Width;
+	size_t m_Height;
 };
 
 
@@ -97,7 +97,7 @@ class CCamera
 		 * @param intermediates determines how many intermediate distance planes should
 		 * be hinted at between the near and far planes
 		 */
-		void Render(uint intermediates = 0) const;
+		void Render(int intermediates = 0) const;
 		
 	public:
 		// This is the orientation matrix. The inverse of this
