+++ source/lib/file/common/file_loader.h
@@ -5,7 +5,7 @@ struct IFileLoader
 {
 	virtual ~IFileLoader();
 
-	virtual unsigned Precedence() const = 0;
+	virtual size_t Precedence() const = 0;
 	virtual char LocationCode() const = 0;
 
 	virtual LibError Load(const std::string& name, shared_ptr<u8> buf, size_t size) const = 0;
